Reformed the main playlist view to use list tiles instead of the grid view tiles
